// shared/types/types.ts
export const allowedFileTypes = ["image/jpeg", "image/jpg", "image/png", "image/webp"];

export enum Colors {
    Green = "bg-green-500/10 text-green-500",
    Yellow = "bg-yellow-500/10 text-yellow-500",
    Red = "bg-red-500/10 text-red-500",
    Blue = "bg-blue-500/10 text-blue-500",
    Purple = "bg-purple-500/10 text-purple-500",
    Gray = "bg-gray-500/10 text-gray-400",
    Pink = "bg-pink-500/10 text-pink-500", 
}

export enum HexColors {
    Green = "#22c55e",    // Tailwind text-green-500
    Yellow = "#eab308",   // Tailwind text-yellow-500
    Red = "#ef4444",      // Tailwind text-red-500
    Blue = "#3b82f6",     // Tailwind text-blue-500
    Purple = "#8b5cf6",   // Tailwind text-purple-500
    Gray = "#9ca3af",     // Tailwind text-gray-400
    Pink = "#ec4899",     // Tailwind text-pink-500 
}